touche pas maryani

+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ncolliau <ncolliau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2014/11/04 13:57:05 by ncolliau          #+#    #+#             */
-/*   Updated: 2014/11/22 11:46:53 by ncolliau         ###   ########.fr       */
+/*   Updated: 2014/12/02 13:54:33 by ncolliau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ char				**ft_strsplit(char const *s, char c);
 char				*ft_strstr(const char *s1, const char *s2);
 char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strtrim(char const *s);
-int					ft_tabstrlen(char **tabstr);
+size_t				ft_tabstrlen(char **tabstr);
 int					ft_tolower(int c);
 int					ft_toupper(int c);
